Client component work

+++ client/src/main/scala/models/component/StaticImage.scala
@@ -4,8 +4,8 @@ import com.definitelyscala.phaserce.{Game, Group, Image}
 import com.definitelyscala.phasercepixi.Texture
 import services.map.MapService
 
-case class StaticImage(game: Game, group: Group, override val name: String, x: Double, y: Double, tex: Texture) extends BaseComponent {
-  val image = new Image(game = game, x = x, y = y, key = tex)
+case class StaticImage(game: Game, group: Group, override val name: String, x: Int, y: Int, tex: Texture) extends BaseComponent {
+  val image = new Image(game = game, x = x.toDouble, y = y.toDouble, key = tex)
   image.name = name
   image.scale = MapService.scalePoint
   group.add(image)
